requirements.txt
anthropic>=0.39.0
supabase>=2.4.0
tavily-python>=0.3.0
apscheduler>=3.10.4
fastapi>=0.110.0
uvicorn>=0.27.1
resend>=0.7.0
python-dotenv>=1.0.1
pydantic>=2.6.4
requests>=2.31.0
